Format correlation statistics

With format_corr() you can format correlation statistics generated from cor.test() output. This is now an internal function superceded by format_stats(), which we recommend using instead.

Usage

format_corr(x, digits, pdigits, pzero, full, italics, type, ...)

Arguments

x

An htest object from cor.test(), t.test(), or wilcox.test().

digits

Number of digits after the decimal for means, confidence intervals, and test statistics.

pdigits

Number of digits after the decimal for p-values, ranging between 1-5 (also controls cutoff for small p-values).

pzero

Logical value (default = FALSE) for whether to include leading zero for p-values.

full

Logical value (default = TRUE) for whether to include means and confidence intervals or just test statistic and p-value.

italics

Logical value (default = TRUE) for whether p label should be italicized.

type

Type of formatting ("md" = markdown, "latex" = LaTeX).

...

Additional arguments passed to methods.

Value

A character string of statistical information formatted in Markdown or LaTeX.

See also

Examples

# format_stats(cor.test(mtcars$mpg, mtcars$cyl))